refactor(frontend): extract SearchPanel into modular components
